made the question tab
I coded the jsx template and CSS for the questionTab component.

A few stylistic changes. I should include the color of each subject each place it is mentioned. That is, include it where the subject title is on the practice page. Also make it so the currently selected object in the dropdown has the color. Also make the tags for the questions have the color of the tag if available.